Add files via upload
This programme applies the hough transform to an video feed from a webcam and outputs its angle, distance, and a recorded video as an avi file. The value of the threshold of the hough transform and the Carny Edge detection must be adjusted and tuned for potentially more accurate results.